test with data

+++ grey_wolf_optimizer.h
@@ -6,6 +6,6 @@
 namespace optimization
 {
 	//enum class function;
-	solution grey_wolf_optimizer(function f, int max_number_of_evaluations,int number_of_agents, float left_bound, float right_bound,int dimension);
+	solution grey_wolf_optimizer(function f, int max_number_of_evaluations,int number_of_agents, double left_bound, double right_bound,int dimension);
 }
