+++ utils.cpp
@@ -6,17 +6,17 @@ using namespace std;
 
 namespace optimization
 {
-	float get_random(float left_bound, float right_bound)
+	double get_random(double left_bound, double right_bound)
 	{
 		static random_device rd;
 		static mt19937 gen(rd());
-		uniform_real_distribution<float> dist(left_bound,right_bound);
+		uniform_real_distribution<double> dist(left_bound,right_bound);
 		return dist(gen);
 	}
 
-	vector< vector<float> >  get_initial_positions(float left_bound, float right_bound, int dimension, int number_of_agents)
+	vector< vector<double> >  get_initial_positions(double left_bound, double right_bound, int dimension, int number_of_agents)
 	{
-		vector< vector<float> > positions(number_of_agents, vector<float>(dimension,0.));
+		vector< vector<double> > positions(number_of_agents, vector<double>(dimension,0.));
 		for(auto &agent : positions)
 		{
 			for(auto &coord : agent)
@@ -27,7 +27,7 @@ namespace optimization
 		return positions;
 	}
 
-	vector< vector<float> > &clip_positions(vector<vector<float>> &positions, float left_bound, float right_bound)
+	vector< vector<double> > &clip_positions(vector<vector<double>> &positions, double left_bound, double right_bound)
 	{
 		for (auto &agent : positions)
     	    for (auto &coord : agent)
